Hd-rum-translator: decompress only if transcoding
The rationale is that hd-rum-translator can be used as a packet
reflector as well, in which case it is capable process also different
packet types than video. In this case, if we do not require transcoding
ports, just disable additional video processing. This has also advantage
for video, if we use only packet reflector - no video decompression is
done and thus it reduces resourcees usage.