GitHub CI: replace cache key counters
GitHub actually uses a bit complicated pattern matching for the keys:
https://docs.github.com/en/actions/guides/caching-dependencies-to-speed-up-workflows#using-contexts-to-create-cache-keys

Namely if exact hit is not found, the key is used as a prefix which
could cause using 'ffmpeg-cache-2' for 'ffmpeg-cache'. In addition
steps.<id>.outputs.cache-hit is set to false if the match is not exact
but the cache is actually restored from inexact match.

+ use ${{ runner.os }} as part of the cache key (currently not needed
but we may use the same key for different platform caches)