audio_frame2: make copy ctor + assign deleted
Make copy constructor and assignment operator explicitly deleted - until
now it has been deleted implicitly so this is just to ensure that this
is actually a desired behavior. The operation in the internal pipeline
should be copy-free (af2 is used only there). If really needed, some
explicit function for copy could be written.