multiplier disp: Fix segfault on display_done
If a display does not need the mainloop, display_run() just spawns a new
thread and immediately returns. This means that the 'disp_thread' thread
isn't the actual display thread and finishes far sooner than the display.
As a result, display_done() was called while the display was still running.