GL: fixed a race condition
Make complete initialization in _init function from now - _run must be
run from main thread now (even if not, the context may be detached and
reattached in a thread.

A race condition occurs when poison pill is passed and the content is
not fully initialized (_run function). Then putf function doesn't mark
the window for deletion, because it has not been created yet and as a
result, the run loop keeps running.