Makefile: Call GUI QMAKE with original C(XX)FLAGS
When a variable is defined in the environment, it also gets modified
and exported for all commands and sub-makes in the Makefile
(https://web.mit.edu/gnu/doc/html/make_6.html#SEC66).

This means that when building with custom C(XX)FLAGS (such as with
Archlinux makepkg), the flag "-DHAVE_CONFIG_H" which absence  is used to
detect out-of-tree building (in platform_pipe.cpp) got also wrongly
passed to qmake. This caused linking to fail due to undefined symbols,
which are only present in UltraGrid in-tree-builds.