setvbuf: set size=BUFSIZ
The value 0 is mentioned in man setvbuf(3) and indeed glibc
implements setlinebuf as:

    _IO_setvbuf (stream, NULL, 1, 0);

However, except the mention in manual page, this extension doesn't seem
to be anywhere mentioned (not clear if valid with _IOFBF) and it is not
widespread except glibc, since [1] forbids that. Also using _IOLBF+0
behaves in the same way as _IOLBF+BUFSIZ (buffers BUFSIZ bytes).

[1] https://www.ibm.com/docs/en/i/7.1?topic=functions-setvbuf-control-buffering