GL: fixed displaying of R10k width % 64 != 0
R10k line is padded to 256 B (as DeckLinks requires) so widths not
divisible by 64 pixels were not displayed correctly.

Not sure if AJA holds also this convetion, it is not mentioned in docs
for NTV2_FBF_10BIT_DPX. However, SDI formats' line widths are generally
divisible by 64 pixels so it is not relevant there.

Steps to reproduce the problem were:

    uv -t testcard:size=48x32:codec=R10k -d gl