jack: Do not automatically start jack server in probe
Jack does not wait for the server to exit completely on client_close(),
which causes a race condition when running uv with `--capabilities`:
when the jack playback device is probed right after probing the jack
capture device, jack still sees the terminating server and tries to
connect to it unsuccessfully 5 times in a row.

This changes reduces the time it takes for --capabilities to run by ~7
seconds, greately reducing the GUI startup time.