lavd: print selected pixfmt as INFO
Decoder may attempt first pixelformat (eg. nv12) that won't be used in
the end because it changes its mind to eg. p010. _But_, the display may
have been already reconfigured to v210 which would cause swscale
conversion for nv12 (but not p010), thus it would seem to be like a
problem in output because the swscale information is shown by default
but subsequent pixel format change isn't.