lavd: remove swscale notice info
User will be informed about swscale convert in
lavc_sws_convert_reconfigure() when appropriate.

The removed message in get_format callback was misleading because decode
may first pick eg. nv12 and after that decides to yuv444p16le, for which
there is a conversion, without decoding a frame in nv12.