libavcodec: lavc_compare_convs - fixed ordering
Fixed wrong ordering in rare cases, eg.:

1. R10k->UYVY->yuv444p
2. R10k->UYVY->yuv422p

For both conversion chain, the same property 10b 4:2:2 is selected. But
as a tie-break, R10k is compared with resulting AVPixelFormat's
properties, winning #1 because seemingly keeping 4:4:4 subsampling, not
reflecting the degradation in chain.

To fix it up, use for the second comparison the minimum of src pixfmt
desc and the conversions' chain (for both are the same as mentioned in
previous paragraph).

+ improved debugging (in debug2, print comparison results & print
  property of the uv->uv-only conversions)