lavd: delay HEVC decoding up to first VPS
Begin decoding of HEVC stream after a frame starting with VPS NALU has
been seen. This improves the behavior in a sense that there is not an
initial flood of decoding errors until first IDR frame is received.

This change is analogous to the one that is already present for H.264
(using SPS NALU).

In case of problems, check if it holds that VPS NAL is always first
(seems to be the case but not sure if it is mandatory; but similarly
this presumption holds for H.264 and there has not be any counterexample
until now).