timespec_get: added compat for macOS
+ added utility function for struct timespec

It is actually more efficient to compute without using multipliciations
and divisions. It is also more readable. Also, C11 concurrency library
uses struct timespec (for timeouts etc.) as well as pthread funcitions.
So it is perhaps better to let developer decide if wants to use single
digit time_ns_t or struct timespec.

+ use it in sdl2 and caca displays