decklink disp.: no-low-latency fixes + deprecate
This fixes f7cc336ca - actually this change only broke the scheduled
playback, without affecting the low-latency mode (where i is always 0).
On contrary, for scheduled playback, where non-blocking is used by
default, the comparison is with 0 instead of 2 as it used to be, leading
to frequent drops.

This also reverts 00801fa3 - scheduling as currently implemented
produces last frames (actually this is not an issue - our workflow is
different than scheduled playback of steady media, that should actually
schedule next frame in completion callback).

Most importantly, deprecate the scheduled mode - not sure if used but if
someone does, we would perhaps need to know its use-case.