sdl_mixer: add as candidate root also /usr
When on Linux, UG might have been installed either to /usr/local prefix,
in which case neither of hard-codec suffixes would match system
SoundFont. Or UG may be run from within build directory directly, in
which case even less likely system sound font could be found.

So add /usr as a fallback (note that it seems to be needed specifically
for Arch Linux, in Ubuntu, Mix_GetSoundFonts() finds the path to system
sound font, so guessing the path is not needed).