Deck disp.: audio oveflow details only if verbose
Print overflow details (nr of samples written/dropped/buffered) only in
verbose mode. The point is that those values will differ and if there is
a moment where overflow occurs, it is better to colligate the overflow
messages into one and only to see the details in verbose mode, otherwise
the output could be flooded with the messages.