GL: deprecate fixed_size
The name is needlessly verbose - it is quite clean, that size of the
window will be fixed (on the other hand, it is still resizable), it just
leaks a semantics that the GL window is by default adjusted to video
size.

We can use simply "size=" parameter. Also deprecated "fixed_size"
without parameters - it doesn't have direct counterpart but it
basically works as "size=512x512" because it fixes to first "video",
which is currently the splashscreen (at the time of writing it used to
be really the first incoming video but is seems there is a little use
case for that - CoUnsil used that at a time).

+ optinally a window position can be specified