RTP: send synchronized A/V TS where possible
Send same RTP timestamps for audio/video frames captured at the same
time.

RTP specification allows this (although doesn't require - according to
the spec, the RTP timestamps should colaesce with RTCP NTP
correspondence). We slightly modify that in a way that we use for both
audio and video 90 kHz clock resolution, so if the start point is the
same (random), the RTP timestamps will be exactly the same for related
data.

Implemented for vidcap DeckLink and testcard.