video_decoder: fixed display reconf on probe
Display should not be reconfigured when not needed, otherwise it will be
reconfigured twice even when received single compressed stream (because
reconf is run twice here - once as probe and second real).

Therefore the display description is stored in decoder->display_desc and
compared with the actual desc. But it doesn't work with compresed stream,
because display_desc.color_spec is set to the network desc and rewritten
to the correct uncompressed color-spec just after the if. So eg.:

- stored desc is "1920x1080 @50.00i, codec UYVY"; **but** compared to:
- actual "1920x1080 @50.00i, codec MJPEG"