DeckLink cap.: allow 32/64 channs unconditionally
It doesn't seem to be restricted by API version, eg. on an Intensity Pro,
16 works for both SDK 11.6 and 12 (second advertises 32/64 channels in
SDK doc, first not), although it has only 2 channels (attrs mach/aach/iach).
32 channels is refused by both SDK versions, so doesn't necessarily need
to be related to version.

Since this setting is on the sender, we do not need  to check the validity
since if invalid, it would fail later, anyways.