to_lavc_vid_conv: added rgb_to_yuv444p
For following command, deduced conversion is to 10-bit YUV:

    $ uv -t testcard:c=RGB -c libavcodec:enc=libx264 -d gl
    [to_lavc_vid_conv] converting RGB to yuv444p10le over R12L

(and over R10k for RGBA) which is correct, because we don't have any
8-bit YUV pixfmt keeping 4:4:4 subsampling.

But this is quite ineffective because the conversions are more expensive
and we are needlessly compressing 10-bit YUV instead of 8-bit.

Thus (as we don't have any UG 8-bit YUV444 pixfmt) the rgb_to_yuv444p
conversion was added.