DeckLink cap.: don't capture audio if nosig
affects c1a1f83b, 64ecc161

Now audio frames are discarded unless if video signal is not detected,
this doesn't affect situations when just the audio frame is passed.

This refers to GH-347, because it drops the audio frames that are produced
prior signal is detected. Calling IDeckLinkInput::EnableVideoInput restart
timestamps.