hd-rum-transcode: port val check updates
1. really check non-negativity - until now, the parsed string was
dirrectly assigned to ushort, effectively disabling the check ((short)-1
== 65535)
2. check also upper bound to be less than USHRT_MAX
3. **allow** 0 as a port - supposedly this is a relict when atoi() was
used to parse the port, so that it was supposed to be error in
conversion. There may be, however, legitimate reasons to use 0 as a port
to bind random port (although UG doesn't print it so the user would need
to obtain its value eg. with iproute2 if required). Just print a warn.