CI mac: switch to macOS 12 Monterey
Homebrew builds now dependencies from sources on unsupported macOSes
(which is 11 now), which makes it slow. Relevant snippet from brew output:
```
Warning: You are using macOS 11.
We (and Apple) do not provide support for this old version.
It is expected behaviour that some formulae will fail to build in this old version.
It is expected behaviour that Homebrew will be buggy and slow.
```