GL: try to fix v210 shader
refer to GH-359

It is expected that the value should been slightly above a pixel index,
eg. 0.0001 for first and 1919.0001 for the last in FHD.

However, using round (but not floor!) before converting the value to
int seem to help, which would imply that the actual pixel value might
have fallen below the actual pixel value. But in that case, replacing
0.4999 for 0.5 should not have helped, but it did (reverting the commit
c2437894f).