vrxtx/rtsp subsessions: do not send msgs on exit
When destroying the state, ServerMediaSubsession::deleteStream callbacks
are called but we are already deconfiguring so do not send messages to
change destination (eg. audio is already destroyed so the message wont
be delivered causing warnings like:
```
Receiver audio.sender does not exist.
Receiver audio.sender does not exist.
Warning: Message queue not empty!
```

Currently this is just a dirty fix, rtsp_serv::watch would be better to
pass.