gl_vdpau: Disable vdpau extension checking
glGetString(GL_EXTENSIONS) is not supported in GL 3.3, so this would
make it impossible to use vdpau from the panorama displays.

The GL displays performs it's own check, so this was a duplicate anyways

The gl_conversions interface for getting supported codec list does not
perform any checks yet, however it seems unlikely that a GPU could
successfuly decode to HW_VDPAU but lack the GL interop extension at the
same time.