dshow: print mode pName instead of ug name
NV12 will be mapped to UYVY, so instead of NV12, there will be
misleadingly written UYVY.

Actually, the FourCC stored in &pSubtype->Data could be used, because
it mostly match. Chosen approach is, however, more conservative if the
name doesn't match, because the "old-style" init with the codec name
requires the mapping back from name to subtype and there doesn't seem
to be a way to find a subtype from FourCC in genral.