vcap/dshow: print streaming/grab type
Instead of getting full format from vidcap_dshow_get_video_desc(), which
translates the source pixel format to UG one, use the FourCC directly. Do
write just the format, the same for grabber and then the complete output
format (after eventual UG conversion). Except of the format, every other
parameters should be currently the same.

The previous approuach didn't work when there was either no VFW->UG
mapping (I420 etc.) or there  was not 1:1 mapping, eg. NV12 was displayed
as UYVY because the conversion is such, which might be misleading.