JACK loader [mac]: try also lib in /usr/local
/usr/local/lib is default location for Homebrew on Intel macs, which is,
however, searched by default. So it is actually not needed on Intels.

It is, however, also the path of the JACK library installed by the
official installer, which uses the universal binary (there is Intel-only
installer, too). With that universal library, the x86_64 build using
Rosetta2 will work with JACK also on ARM64 macs.