vcompress/lavc SVT: preset for HEVC > UHD@60
The SVT HEVC compression has (for UG) an ugly feature that it caches
internally in some cases a huge amount of frames, leading to increased
latency and also memory consumption. The first problem is dealt with
at least with the warning (commit e3f64d56, 2024-04-12).

if running
`uv -t testcard:c=RG48:s=3840x2160:fps=240 -c lavc:enc=libsvt_hevc`
with default preset (7) the latency grew up to some 430 frames, triggering
the assertion in store_metadata().

Setting the preset (for the above command - it is size/FPS depending)
to 11 shrinks the amount of frames cached to some 40 frames, that seem
more acceptable.