CoreAudio: accept name for default devices
The default device name was not included in the list of device for
comparison, because the direction was not given. But we know the direction
both in capture and playback, so we can add it so that the default device
is recognized by name.

This also makes audio_ca_probe() more clear as a direction is always
capture or playback (-1 or 1).

\+ added some doxygen documentation

refer to GH-385