{vcap,vdisp}/deck: added "simplex" option
Added option "simplex" in a fasshion of "half-duplex" and "full-duplex",
which are informal shortcuts for device profiles.

Here, simplex refers to bmdProfileOneSubDeviceHalfDuplex (1dhd) for 8K
Pro, which basically sets device to a mode that is capable to use all
four connectors for the signal, effectively allowing quad-link input
(eg. 8K).