deck: topological ID sort fix
Limit lifetime of the returned map to the ranged for scope - the map
actually contains references to IDeckLink devices with deleters.

In probes and helps, it meant that release was called at the block
end, which means in this case that the release occurs right after
decklink_uninitalize(), which is certainly incorrect.

It seems that isn't a big problem in inits - it would be a problem just for
devices in the map after the initialized one (previous are unmanaged
and released) and if init fails (decklink_uninitalize will be called
before the block ends in this case).