video_rxtx/rtsp: use port=number for RTSP port spec
Prefer `-x rtsp:port=<P>` instead of `port:<P>`. The former one is a bit
inconsistend with other stuff and would also cause problems in future if
more options were added.

Keep the compat for the old syntax.

Fixed the usage, which was not correct anyways (the '=' should have been ':').