configure [Win]: add CUDA_LIB to LIBS
Add CUDA_LIB to LIBS unconditionally as it is already done for other
platforms except Win.

In Windows, the CUDA wrapper was built outside UltraGrid (as a DLL) but this
is no longer true since the commit a4c22c93f (2024-03-05) so if there is no
CUDA features but still CUDA enabled (as with `--disable-all --enable-cuda`),
the libs were not added for the cuda_wrapper.