vcf/logo: CID 467080 fix
Fixed the Coverity complaint by keeping the behavior of "doing nothing" with
the frame while passing it unchanged. It may also return nothing, as it was
doing for the short time - this seems to be inferior in cases when the input
changes from supported to unsupported, this would cut-off the picture eniterely,
not just the logo.