vdisp/sdl2: call SDL_VideoInit before SDL_Init
In order to the initialize the option `driver=` successfully the video
subsystem must be called with the driver name (SDL_VideoInit), not with
SDL_Init[SubSystem].

This allows eg. `-d sdl:driver=KMSDRM` to be used.

Also use the driver in show_help (for `-d sdl:driver=KMSDRM:help`),
otherwise will SDL_Init() in help complained that "x11 not available".

refers to GH-412