vdec/lavd: do not print JPEG warning
The JPEG is either BT.709 (lavc-encoded, YCbCr encoded by GPUJPEG)
or Adobe RGB (if encoded by GPUJPEG). Both are processed correctly
by the lavd decoder - BT.709 is assumed if YCbCr, RGB is decoded
to GBRP (so either no conversion to RGB PF or default handling
if the output PF is YCbCr).