audio codec: if received no data, pass it further
This is an exception to the normal behavior used by the passive audio
mixer participant, that sends empty packets to request receiving the data.

It doesn't affect the normal UG behavior, because the situation when
nonzero_channels=0 doesn't occur otherwise. Either we receive at least
some packet from a frame and if not, the decoding doesn't take place.