AppRun: if update is due, touch first
If ULTRAGRID_AUTOUPDATE is set and the update is due, touch the AppImage
first. This is convenient if user wants to skip the update eg. by pressing
Ctrl-C, next run will not enforce the update becauuse the date of the file
is reset. User can run then run the AppImage by issuing the command again
without the update.

+ notice that updating because of ULTRAGRID_AUTOUPDATE is set
+ rename handle_updates to handle_autopudates (the normal update by -u is
not handled by that function at all)