configure.ac: prefer 2 over 3 for now
but not if sdl2 is SDL3 sdl2_compat - this won't bring any advantage

There is still some issues with 3 that was not with v2 - see sdl3.c and
sdl_mixer.c heading. Also, there is still no stable SDL_ttf release,
and even not a preview v3 build of SDL_mixer.