vcomp/lavc: accept no_intra_refresh
and no_interlaced_dct

The original option names are prefixed with disable_ to disable. But if
user (mistakenly) uses "no_XY", it actually did the opposite. So rewrite
the condition to accept both variants for now. The "no_" prefix may be
more natural and it might be switched to in future.

Also accept "no_header_inserter" aside the original "header_inserter=no"
- the original syntax is inconsistent with the other options above - this
could be unified later.

+ removed duplicite line