video_compress: module state changes to void *
The internal state of compression module is now simply (void *) as in
other modules. The (module *)-typed vesion was actually unnecessary
because except lavc, no module reads messages. Also the idea of universal
API with struct module didn't make it in the end.

This complements the commit 0557730 that did simlarly for external
vcompress API.