sdl1: do not register q to keycontrol
In order not to be pressed by accident - other SW displays disable 'q'
as well.

Also, the call was incorrect - the message with the key would receive
root module because the first argument is the receiver of the key event,
which should be the sdl display mod. Also the events were not handled in
check_message event.