utils/opencl: set CL_TAERGET_OPENCL_VERSION=120
CL headers issue warning when not set and default to eg. 300 (3.0).

Not sure if this influences just this code but it seems that this may just
limit API availability provided by hdrs? So we can get along with !.2 (for
CL_DEVICE_TYPE_CUSTOM).

We currently only do device listing to pass further (Cmpto J2K codec) so
AFAIK we don't have initialization of the device so this shouldn't matter
(at most we'll allow device with unsupported OpenCL version but perhaps
cmpto codec could catch this)