deltacast_dvi: improve code flow for HDMI boards
The original sample code uses cascadding if-ladder evaluating Result. The
derived code was just slightly derived but it was maybe even worse because
the error handling (message) was coupled in one if/else with subsequent
API call.

This is mostly a refactor but:
1. two-times VHD_DV_SP_REFRESH_RATE removed
2. perhaps fixed misaligned error message prints to wrong calls, eg. when
board type was not HDMI, the error was print even though the previous
call didn't fail (but continued, anyways)