to_lavc_vid_conv: watch_pixfmt_degrade improved
compare just the in codec and the MIN(intermediate, out_pixfmt)

to avoid missleading message
`[to_lavc_vid_conv] conversion is reducing subsampling from 4444 to 4440`
as eg. in:
<https://github.com/CESNET/UltraGrid/issues/480>

if converting R12L to bgra over RGBA - the alpha channel is present just
in the intermediate but both ends have 444 without alpha channel.