vcomp/lavc: print the list if no codec found
print_pix_fmts() is called `if (pix_fmt == AV_PIX_FMT_NONE || log_level
>= LOG_LEVEL_VERBOSE)`

But in the former case (if no pixel format was found) and log level <
VERBOSE, nothihg was actually print, because the function used VERBOSE
log level to the printout.

Use the current log_level to be print unconditionally (if called).